pyproject.toml
[project]
name = "anthropic-computer-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.37.1",
    "pillow>=11.0.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project.scripts]
main = "anthropic_computer_use.main:main"
